Inline the isOsArchitectureCompatibleWith utility to the OS model
